[#18] Refator code, change destination, update for EOL check
