--- client/src/pages/Home.css
@@ -1,259 +1,0 @@
-/* Home Container */
-.home-container {
-    max-width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    flex-direction: column;
-    background-color: rgba(7, 82, 126, 0.85);
-}
-
-/* Introduce Section */
-.introduce-box {
-    width: 900px;
-    height: 450px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.introduce-box-left h4 {
-    padding: 0px;
-    margin: 0px;
-    font-size: 38px;
-    font-weight: 600;
-    line-height: 42px;
-    margin-bottom: 20px;
-    color: white;
-}
-
-.introduce-box-left p {
-    padding: 0px;
-    margin: 0px;
-    font-size: 18px;
-    line-height: 20px;
-    margin-bottom: 20px;
-    color: rgb(204, 204, 204);
-}
-
-.introduce-box-left button {
-    background-color: yellow;
-    font-weight: 600;
-    padding: 10px;
-    font-size: 16px;
-    border: none;
-    border-radius: 4px;
-}
-
-.introduce-box-left button:hover {
-    cursor: pointer;
-    text-decoration: underline;
-    background-color: yellowgreen;
-}
-
-.introduce-box-right img {
-    padding: 0px;
-    margin: 0px;
-    width: 300px;
-}
-
-/* Information Section */
-.information-box {
-    width: 100%;
-    background-color: rgb(240, 244, 255);
-    background-size: cover;
-    background-repeat: no-repeat;
-    display: flex;
-    justify-content: center;
-    padding: 20px 0;
-}
-
-.information {
-    width: 980px;
-    background-color: white;
-}
-
-.information h4 {
-    padding: 10px;
-    margin: 0px;
-    text-align: center;
-    font-size: 30px;
-}
-
-.information p {
-    padding: 0px 30px;
-    margin: 0px;
-    text-align: center;
-    font-size: 18px;
-    line-height: 1.5;
-}
-
-/* Info Cards */
-.info-wrapper {
-    margin: 0 auto;
-    text-align: center;
-    margin-bottom: 30px;
-}
-
-.info-title {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 20px;
-    color: #333;
-}
-
-.info-description {
-    font-size: 16px;
-    color: #555;
-    margin-bottom: 40px;
-    line-height: 1.5;
-}
-
-.info-features-container {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-}
-
-.info-feature-card {
-    background: #fff;
-    border-radius: 10px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-    padding: 20px;
-    margin: 5px;
-    text-align: center;
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.info-feature-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.15);
-}
-
-.info-icon {
-    width: 70px;
-    height: 70px;
-    margin: 0 auto 20px;
-    border-radius: 50%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.info-icon img {
-    width: 40px;
-    height: 40px;
-}
-
-.info-feature-title {
-    font-size: 18px;
-    font-weight: bold;
-    margin: 10px 0;
-    color: #333;
-}
-
-.info-feature-description {
-    font-size: 14px;
-    color: #555;
-    margin: 0;
-    line-height: 1.5;
-}
-
-/* Responsive Design */
-@media (max-width: 920px) {
-    .introduce-box {
-        width: 100%;
-        flex-direction: column;
-        text-align: center;
-        justify-content: center;
-        align-items: center;
-        height: fit-content;
-    }
-
-    .introduce-box-left {
-        margin: 20px 10px;
-    }
-
-    .introduce-box-left h4 {
-        font-size: 26px;
-    }
-
-    .introduce-box-left p {
-        font-size: 16px;
-    }
-
-    .introduce-box-left button {
-        font-size: 14px;
-        padding: 8px;
-    }
-
-    .introduce-box-right {
-        display: none;
-    }
-
-    .info-title {
-        font-size: 22px;
-    }
-
-    .info-description {
-        font-size: 15px;
-    }
-
-    .info-feature-title {
-        font-size: 16px;
-    }
-
-    .info-feature-description {
-        font-size: 13px;
-    }
-
-    .info-icon {
-        width: 60px;
-        height: 60px;
-    }
-
-    .info-icon img {
-        width: 30px;
-        height: 30px;
-    }
-}
-
-@media (max-width: 480px) {
-    .introduce-box-left h4 {
-        font-size: 22px;
-    }
-
-    .introduce-box-left p {
-        font-size: 14px;
-    }
-
-    .introduce-box-left button {
-        font-size: 14px;
-        padding: 8px;
-    }
-
-    .info-title {
-        font-size: 20px;
-    }
-
-    .info-description {
-        font-size: 14px;
-    }
-
-    .info-feature-title {
-        font-size: 15px;
-    }
-
-    .info-feature-description {
-        font-size: 12px;
-    }
-
-    .info-icon {
-        width: 50px;
-        height: 50px;
-    }
-
-    .info-icon img {
-        width: 25px;
-        height: 25px;
-    }
-}
