--- client/src/pages/Layout.css
@@ -1,270 +1,0 @@
-/* General Layout */
-.layout-container {
-    min-height: 98vh;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-.layout-container main {
-    width: 100%;
-}
-
-/* Header Section */
-.layout-container header {
-    width: 100%;
-    padding: 0;
-    margin: 0;
-    background-color: #002244;
-    display: flex;
-    justify-content: center;
-}
-
-.header-content {
-    width: 1200px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background-color: #002244;
-}
-
-.page-icon {
-    max-width: auto;
-    max-height: 60px;
-    padding: 5px 0px;
-    cursor: pointer;
-}
-
-.page-icon:hover {
-    cursor: pointer;
-}
-
-/* Navigation Bar */
-.layout-container header ul {
-    list-style: none;
-    display: flex;
-    padding: 0px;
-    margin: 0px;
-}
-
-.layout-container header ul li {
-    margin: 0px 18px;
-    font-size: 20px;
-    font-weight: 500;
-    color: #00ff80;
-    padding: 5px;
-    border-radius: 8px;
-    text-wrap: nowrap;
-}
-
-.layout-container header ul li:hover {
-    cursor: pointer;
-    color: #d0ff00;
-    scale: 1.1;
-    text-decoration: underline;
-}
-
-#login-btn {
-    color: #fbff00;
-}
-
-#login-btn:hover {
-    color: #00ff80;
-}
-
-/* Menu Toggle Button */
-.menu-toggle {
-    display: none;
-    background: none;
-    border: none;
-    color: #00ff80;
-    font-size: 24px;
-    cursor: pointer;
-    max-height: 32px;
-    margin-right: 10px;
-}
-
-/* Footer Section */
-.layout-container footer {
-    width: 100%;
-    padding: 0;
-    margin: 0;
-    background-color: #002244;
-    display: flex;
-    justify-content: center;
-}
-
-.footer-container {
-    width: 100%;
-    max-width: 1200px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    background-color: #002244;
-}
-
-.section-container {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    margin-bottom: 30px;
-    width: 100%;
-}
-
-.footer-section {
-    text-align: center;
-    margin: 0px 10px;
-}
-
-.footer-title {
-    font-size: 18px;
-    font-weight: bold;
-    margin-bottom: 15px;
-    color: #00ff80;
-    position: relative;
-}
-
-.footer-title::after {
-    content: "";
-    display: block;
-    height: 3px;
-    background-color: #00ff80;
-    margin-top: 5px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.footer-icon {
-    width: 90px;
-}
-
-.footer-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.footer-list li a,
-.footer-icon {
-    margin: 5px;
-    padding: 5px;
-    font-size: 14px;
-    color: #bbb;
-    text-decoration: none;
-}
-
-.footer-list li a:hover,
-.footer-icon:hover {
-    cursor: pointer;
-    color: #f6ff7d;
-    border-radius: 8px;
-}
-
-/* Social Media */
-.social-media {
-    display: flex;
-    justify-content: center;
-    gap: 10px;
-    margin-bottom: 20px;
-}
-
-.social-btn {
-    text-transform: uppercase;
-    font-size: 14px;
-    font-weight: bold;
-    text-align: center;
-    padding: 10px 20px;
-    border-radius: 5px;
-    color: #fff;
-    text-decoration: none;
-    display: inline-block;
-}
-
-.social-btn.facebook {
-    background-color: #3b5998;
-}
-
-.social-btn.linkedin {
-    background-color: #0077b5;
-}
-
-.footer-bottom {
-    text-align: center;
-    font-size: 14px;
-    color: #bbb;
-}
-
-.footer-bottom span {
-    color: #ffcc00;
-}
-
-/* Responsive Design */
-@media (max-width: 780px) {
-    .nav-bar {
-        display: none;
-        flex-direction: column;
-        position: absolute;
-        top: 70px;
-        right: 0px;
-        background-color: #002244;
-        box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-        border-radius: 8px;
-        z-index: 1000;
-    }
-
-    .nav-bar.open {
-        display: flex;
-    }
-
-    .nav-bar ul {
-        align-items: center;
-        list-style: none;
-        padding: 0;
-        margin: 0;
-        text-align: right;
-    }
-
-    .nav-bar.open ul li {
-        padding: 5px 20px;
-        color: #00ff80;
-        font-size: 18px;
-    }
-
-    .nav-bar ul li:hover {
-        color: #d0ff00;
-    }
-
-    .menu-toggle {
-        display: block;
-    }
-
-    .nav-bar {
-        display: none;
-    }
-
-    .nav-bar.open {
-        display: flex;
-    }
-
-    .nav-bar ul {
-        flex-direction: column;
-    }
-
-    .nav-bar ul li {
-        margin: 0;
-    }
-
-    .footer-title {
-        font-size: 16px;
-    }
-
-    .footer-list li {
-        font-size: 12px;
-    }
-
-    .social-btn {
-        font-size: 12px;
-    }
-}
