+++ client/src/pages/layout/LayoutFooter.css
@@ -1,0 +1,138 @@
+:root {
+    --color-primary: #002244;
+    --color-accent: #00ff80;
+    --color-hover: #f6ff7d;
+    --color-muted: #bbb;
+    --color-highlight: #ffcc00;
+    --color-facebook: #3b5998;
+    --color-linkedin: #0077b5;
+    --color-white: #fff;
+}
+
+/* Footer Section */
+.layout-container footer {
+    width: 100%;
+    padding: 0;
+    margin: 0;
+    background-color: var(--color-primary);
+    display: flex;
+    justify-content: center;
+}
+
+.footer-container {
+    width: 100%;
+    max-width: 1200px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    background-color: var(--color-primary);
+}
+
+.section-container {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    margin-bottom: 30px;
+    width: 100%;
+}
+
+.footer-section {
+    text-align: center;
+    margin: 0px 10px;
+}
+
+.footer-title {
+    font-size: 18px;
+    font-weight: bold;
+    margin-bottom: 15px;
+    color: var(--color-accent);
+    position: relative;
+}
+
+.footer-title::after {
+    content: "";
+    display: block;
+    height: 3px;
+    background-color: var(--color-accent);
+    margin-top: 5px;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.footer-icon {
+    width: 90px;
+}
+
+.footer-list {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+.footer-list li a,
+.footer-icon {
+    margin: 5px;
+    padding: 5px;
+    font-size: 14px;
+    color: var(--color-muted);
+    text-decoration: none;
+}
+
+.footer-list li a:hover,
+.footer-icon:hover {
+    cursor: pointer;
+    color: var(--color-hover);
+    border-radius: 8px;
+}
+
+/* Social Media */
+.social-media {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+    margin-bottom: 20px;
+}
+
+.social-btn {
+    text-transform: uppercase;
+    font-size: 14px;
+    font-weight: bold;
+    text-align: center;
+    padding: 10px 20px;
+    border-radius: 5px;
+    color: var(--color-white);
+    text-decoration: none;
+    display: inline-block;
+}
+
+.social-btn.facebook {
+    background-color: var(--color-facebook);
+}
+
+.social-btn.linkedin {
+    background-color: var(--color-linkedin);
+}
+
+.footer-bottom {
+    text-align: center;
+    font-size: 14px;
+    color: var(--color-muted);
+}
+
+.footer-bottom span {
+    color: var(--color-highlight);
+}
+
+/* Responsive Design */
+@media (max-width: 780px) {
+    .footer-title {
+        font-size: 16px;
+    }
+
+    .footer-list li {
+        font-size: 12px;
+    }
+
+    .social-btn {
+        font-size: 12px;
+    }
+}
